--- full_stack_python/auth/__init__.py
@@ -1,5 +1,0 @@
-from . models import UserInfo
-
-__all__ = [
-    'UserInfo'
-]
