+++ full_stack_python/blog/__init__.py
@@ -2,7 +2,6 @@ from .add import blog_post_add_page
 from .detail import blog_post_detail_page
 from .edit import blog_post_edit_page
 from .list import blog_post_list_page
-from .model import BlogPostModel
 from .state import BlogPostState
 
 __all__ = [
@@ -10,6 +9,5 @@ __all__ = [
     'blog_post_detail_page',
     'blog_post_edit_page',
     'blog_post_list_page',
-    'BlogPostModel',
     'BlogPostState',
     ]
