+++ full_stack_python/blog/list.py
@@ -2,10 +2,10 @@ import reflex as rx
 
 from .. import navigation
 from ..ui.base import base_page
+from ..models import BlogPostModel
+from . import state
 
-from . import state, model
-
-def blog_post_detail_link(child: rx.Component, post: model.
+def blog_post_detail_link(child: rx.Component, post:
 BlogPostModel):
     if post is None:
         return rx.fragment(child)
@@ -19,7 +19,7 @@ BlogPostModel):
         href=post_detail_url
     )
 
-def blog_post_list_item(post: model.BlogPostModel):
+def blog_post_list_item(post: BlogPostModel):
     return rx.box(
         blog_post_detail_link(
             rx.heading(post.title),
