+++ full_stack_python/blog/state.py
@@ -5,7 +5,7 @@ import reflex as rx
 from sqlmodel import select
 
 from .. import navigation
-from .model import BlogPostModel
+from ..models import BlogPostModel
