+++ full_stack_python/full_stack_python.py
@@ -29,8 +29,10 @@ class State(rx.State):
         
 def index() -> rx.Component:
     # Welcome Page (Index)
+    my_user_obj = SessionState.authenticated_user_info
     my_child = rx.vstack(
             rx.heading(State.label, size="9"),
+            rx.text(my_user_obj.to_string()),
             rx.text(
                 "Get started by editing ",
                 rx.code(f"{config.app_name}/{config.app_name}.py"),
