+++ full_stack_python/models.py
@@ -1,0 +1,80 @@
+from datetime import datetime
+import reflex as rx
+from reflex_local_auth.user import LocalUser
+
+import sqlalchemy
+from sqlmodel import Field, Relationship
+
+from . import utils
+
+class UserInfo(rx.Model, table=True):
+    email: str
+    user_id: int = Field(foreign_key='localuser.id')
+    user: LocalUser | None = Relationship() # LocalUser instance
+    created_at: datetime = Field(
+        default_factory=utils.timing.get_utc_now,
+        sa_type=sqlalchemy.DateTime(timezone=True),
+        sa_column_kwargs={
+            'server_default': sqlalchemy.func.now()
+        },
+        nullable=False
+    )
+    updated_at: datetime = Field(
+        default_factory=utils.timing.get_utc_now,
+        sa_type=sqlalchemy.DateTime(timezone=True),
+        sa_column_kwargs={
+            'onupdate': sqlalchemy.func.now(),
+            'server_default': sqlalchemy.func.now()
+        },
+        nullable=False
+    )
+
+
+class BlogPostModel(rx.Model, table=True):
+    # user
+    # id: int -> primary key
+    title: str
+    content: str
+    created_at: datetime = Field(
+        default_factory=utils.timing.get_utc_now,
+        sa_type=sqlalchemy.DateTime(timezone=True),
+        sa_column_kwargs={
+            'server_default': sqlalchemy.func.now()
+        },
+        nullable=False
+    )
+    updated_at: datetime = Field(
+        default_factory=utils.timing.get_utc_now,
+        sa_type=sqlalchemy.DateTime(timezone=True),
+        sa_column_kwargs={
+            'onupdate': sqlalchemy.func.now(),
+            'server_default': sqlalchemy.func.now()
+        },
+        nullable=False
+    )
+    publish_active: bool = False
+    publish_date: datetime = Field(
+        default=None,
+        sa_type=sqlalchemy.DateTime(timezone=True),
+        sa_column_kwargs={},
+        nullable=True
+    )
+    
+    
+def get_utc_now() -> datetime:
+    return datetime.now(timezone.utc)
+
+class ContactEntryModel(rx.Model, table=True):
+    user_id: int | None = None
+    first_name: str
+    last_name: str | None = None
+    email: str | None = None # = Field(nullable=True)
+    message: str
+    created_at: datetime = Field(
+        default_factory=utils.timing.get_utc_now,
+        sa_type=sqlalchemy.DateTime(timezone=True),
+        sa_column_kwargs={
+            'server_default': sqlalchemy.func.now()
+        },
+        nullable=False
+    )
